Correct previous page of updates to application
There were multiple ways to reach this page but always would return 'dates-of-placement'.
The logic for the forking is in the 'reason for placement' page's 'next' method which
creates different paths for the the 3 reasons ('rotl', 'additional_placement', 'release_following_decision').
Each routes has 1 or more pages in it before reaching 'updates to application' that all paths
meet at.