Add a note
This is a large commit but it contains the full slice of work needed to enable
adding a note on the timeline view.
One thing thing worth noting is the confirmation page for the note. I think this is useful
as there is currently no way of removing a note added in error so we want to make sure that
the note has been checked and confirmed. To do this we need to make two post requests and pass
through the content of the note as a hidden input.