Remove offenceId when setting up applications
This is an internal Delius ID, so isn’t relevant to us. The new
AP to Delus API does not return it, so we need to tweak the
new application journey.

To save all the calls to `getOffences`, we instead pass around
the entire JSON object. If the size of the offences array is
more than one, we redirect to a redesigned select offences screen.