Remove link from in progress applications
If an application is in progress (EG it has the 'started' status) then we don't want
people to be able to click on it and start editing it from the 'All applications' (dashboard)
view. So we pass in a boolean from the view which, when negative,  ensures if an application
isnt linked.