Add multiple 'date of placement' inputs
Users need to be able to add multiple dates of placement on this page.
We could use the MoJ Design system's 'Add Another' component but out of the box this
component isn't accessible and it requires us to reload the same page when a user clicks 'Add
another' with clientside JS turned off. This doesn't work with the form framework we use with
Applications/Assessments and Placement Applications.
Therefore I've copied the approach used in the Contingency Plan Partners Page class and view.
The clientside JS is similar but due to using the GDS Date component on this page it cannot
be genericised enough to be reusable without significantly complicating the code. We don't currently
have any bundling so it is difficult to import clientside JS functions across pages and adding
bundling isn't worth the complications it introduces for the minimal amount of clientside JS
we have.  If, in future, we add a 'Add another' button to another form page I suggest we look at making
the clientside JS reusable across all 3 pages.

For accessibility purposes we render 5 date inputs by default in case clientside JS is off (we don't
expect this to be the case for any users, but need to cover it - just in case). If there
are more than 5 dates then the user will need to either: enable JS or complete another application