Use new endpoint to fetch Placement Applications
Before we were fetching all tasks, but only displaying the Placement
Applications, which was a waste of resources (and also causing
timeouts!). This will still return a lot of data, but less data than
we were requesting before!

Next step is paginating the response, but let’s do one thng at a time