Add 'Enter risk level' page
This page comes at the start of the Apply journey.
If the user states the PiP has a high/very high risk level they will proceed to the exceptional case page.
If they state the PiP has a medium risk level they will proceed to the 'Not eligible' screen.

The tests and page class follow the established Form Pages pattern.