Show multiple dates of placement in list view
Now that a Placement Application (PA) can have multiple dates of placement we need to update
the display of the PAs in the readonly Application view.
We handle all Dates Of Placement as if there are more than of them now but what the users won't
be changed unless there are multiple dates of placement.