Throw error if OASys summary can't be found
We assume that when a summary can't be found it's because the OASys pages haven't been completed as we'd
expect. Users can click through the tabs instead of clicking the 'Save and continue' button at the bottom
of the page which is harder to find.
The result of this is an error on the check your answers screen.