Update endDates page to include new dates
We need to ask for more dates on this page so lets add them.
We also correct the PSS date to PSS end date and PSS start dates as requested by our SME.
There are now a lot (6) dates on the page so we show checkboxes for each one and if
a user selects a date as being relevant then we show a date input.
This means we need an array of selected dates as well as the dates themselves.

In shouldShowContingencyPlanPages we retain the legacy 'pssDate' in case someone is mid-way through an
application as the changes are deployed. This can safely be removed once a suitable amount
of time has elapsed.