Add support for legacy Placement Application data
Before we added support for multiple dates in placement applications, we
had different field names. As some placement applications still have
this data, we need to support both kinds. If `datesOfPlacement` is not
found, then we attempt to fetch from the legacy fields.

I’ve also removed the calls to `addResponseToFormArtifact`, as we’re
already mocking  `pageDataFromApplicationOrAssessment`, so adding the
responses doesn’t actually do anything.