Add Users table
We need to be able to show which users the task can be assigned to as those user's workloads.
This replaces the existing dropdown menu.
It feels a little awkward having to pass the csrfToken through to the table rows function but I
can't see a better way around it.