Refactor shouldShowTableRowsFunction
We need this function in a subsequent commit.
However we will need to construct table rows that take multiple arguments and this function
only allows for a single argument.
We could allow for multiple types using rest parameters but I think it makes sense to have
the consuming function construct the table rows and pass them to shouldShowTableRows. This way
we don't have to worry about the arguments the tableRowFunction takes at all.