Show Offender ‘Status on Application Submission’
All pages that show information about a submitted application have been updated to show the person’s status when the application was submitted, instead of the person’s status at the point in time the page is viewed.

This is achieved by using a new ‘personStatusOnSubmission’ field on the Application model.

The person status tag has also been updated to support showing the ‘Unknown’ type