Add function to calculate adjacent pages for Suitability Assessment task
This function takes the assessment and the current page and returns the next/previous page.
This was necessary as the logic for which pages should be shown was being duplicate for each page in the Suitability Assessment.
For each page added the logic would have to get more complicated.
Now there is a single function which controls which pages should be displayed.