Make responses optional
When we use this function in the new suitabilityAssessmentAdjacentPage function (to be added
in a later commit) it will error as the applications won't have necessarily answered the required questions.
As we are explicit about the responses we expect it is safe to use the `optional`  version
of the function where undefined could be returned instead of erroring if the result cant be found.