Convert assessment test  pages to new Assess page
Previoulsy the data for the assessments was constructed by hard coded args in the assess.cy.ts
file, the assess helper and assessmentData.json.
Whilst the stubbing was simpler it was hard to reason about where the data for the form was coming
from and, most importantly, it was very difficult for assessments to be stateful.
This major rewrite means that all the test data lives in the assessmentData.json and each test
case makes adaptions to this data. The hope is that this will be easier to follow and add subsequent
stateful behaviour.