Add screen to confirm requesting more information
This also adds the creation of the note to the initializer, rather than
reverting to a controller method, which keeps the integration tighter.
If a the user has not already confirmed, and the confirmation field is
set to “yes”, then we create the note and send the user to the next
page (which confirms the note has been set and will be added in the
next commit).

Similiarly, if the confirmation has been set already been set and we’re
looping through the form objects, we set the next page to
'information-received’, so the flow knows that our next step will be
to confirm the information has been received.