Refactor Premises service/view
This was one of the first views added and followed the old pattern of creating parts of the
view in the service.
Since [ADR8](https://github.com/ministryofjustice/hmpps-approved-premises-ui/blob/main/doc/architecture/decisions/0008-data-manipulation-for-views-should-happen-in-util-files.md)
we do manipulation for views in util files, not in services. As we will be updating this
method it makes sense to bring it up to date with the ADR.