Add new method for probationRegions and apAreas
Probation Regions and AP areas have a different type signature to the rest of the Reference data.
Rather than having to handle this in the consuming code it makes sense that they have their
own method.
Probation regions won't be used in this PR as originally though but they may be used in future so lets keep them around.