Supply default reason for late application
Although we show a user an error when they don't enter a reasonForLateApplication some assessments
completed or started before reasonForLateApplication was added will break.
Here we add a default reason so that if a reason isn't supplied we show an error message.
However users will be unable to proceed if they don't supply an answer to this question.