Update possible release types
Now we have two extra release types, we need to additionally show
`paroleDirectedLicence` as an option when the sentence type is
`standardDeterminate`, as well as show `extendedDeterminateLicence`
and `paroleDirectedLicence` if the sentence type is
`extendedDeterminate`.

This makes the logic a bit more complex, so I’ve added `omit` and `pick`
from the `underscore` library to make which release types should be
returned a bit clearer (as destructuring assignements can be a bit of
a mind-melter!) as well as made some tweaks to the types to (hopefully!)
tidy things up.