Add getter for area
If the area property isnt in the body then we want to use the version from Delius.
We need this because the 'area' is a first class field which is sent to the API in
getApplicationData function. This means we need to add both the name and the ID of the Area to
the UserDetailsFromDelius object.