Add queries to `hrefPrefix`
The pagination information was losing the filter queries.
I’ve also made the queries singular, as we’re only searching by
one area/role/qualification etc and it was a bit confusing.

I’ve additionally got rid of the `stubUserFilter` stub method, as
the bulk of the logic was covered in `stubUsers` and uses the
`queryParameters` API withing wiremock, making is easier / more
reliable to match query params.