Ensure that the status filter is retained when another filter is applied
The status filter is applied using tabs. When the user is on a given tab and they click to apply
additional filters to the tab they would not expect the tab to then revert to default (notMatched).
We need to retain the tab filter by passing it through the template as a hidden input.
This means that we need to make the 'get' a 'post' in order to access the body params which
makes navigation a bit more tricky.