Add fieldset legend
As the table is now wrapped in a fieldset, moving the 'Prison case notes' description from the table `caption` element to a new `legend` element which describes the table.

I believe this should provide appropriate detail for the table as it is inside a form group with a description.

This also helps with the placement of the error message, as GOV.UK guidance asks us to place the error after the legend or input label.