Allow placement requests and placement applications to be withdrawn through new controller
This adds the functionality for withdrawing applications, PRs and PAs through the new controller.
We will list out each of the 'withdrawables' for the user to select and then rereoute them
to the appropriate confirmation page (to be updated in a later PR).
We cannot link to a 'show' page for Placement Applications as the page doesnt exist.
The links to withdraw the PA/PR from where they are withdrawn currently have not been updated as
part of this work - this will be done in a subsequent PR.
Bookings require a little more work which will be added in the next commit.