Use AP Area correctly
I made an incorrect assumption that a user’s area already existed in the
session, and it was an ID. Now we’ve actually added it, it’s an object,
so let’s update the tasksController to fetch the ID from the `apArea`
object. I’ve also DRYed up the tests a bit.