Don’t filter section when information not recieved
We’ve had feedback that if a user says they haven’t received feedback
and then subsequently go on to accept the application, there’s a risk
that some important risk information will be missed. This updates the
`getSections` helper to remove the filtering out of this section.

We’ve decided that the pragmatic solution is to remove this, even
though it means the flow may be longer for users.

We still remove the matching information if a user decides to reject
the application, as this is still useful.