Add hidden input to send allocatedFilter back to controller
When a user applies an area filter we need to retain the `allocatedFilter` value, which is set using the tab navigation.
Sharing this as a hidden input value in the form request as in #1412

Move apAreaId used for filtering in integration tests to file level const - as this is used in several tests and actions.

Also disentangle the action to click on the unallocated tab from the assertion `shouldShowUnallocatedTasks`, so that this can be used to assert the right results are shown when directly visiting that tab, or when applying an additional filter.