Add bookings into new Withdrawal flow
Similar to #1440 #1439 we now add bookings to the new withdrawal flow.
As the booking can be withdrawn from the Application now we need to show and hide different inputs for the different roles. Managers will see all the input and other roles won't.
As the date input may not be shown we will instead pass the API todays date if the user doesnt submit a date.
We also update references to 'cancelled' to 'withdrawn'.