Fix unable to filter by all areas
On the Task allocation dashboard and the CRU dashboard we are filtering the table by the users AP Area by default - by setting the users AP Area id as the query value if no value is selected. However as the 'All areas' option in the select is associated with a blank value, this means the user can no longer select All Areas - it will always default to their area.

Instead adding a value of 'all' to the select option 'All areas', and changing this to undefined in the controller.

In order to make this change I have added a default parameter to the nunjucks global `convertObjectsToSelectOptions`, which allows us to pass in an alternative value for the first select item. I believe 'context' cannot be passed in from the nunjucks template, but I can see it has been passed in in some places, so we may want to check these.