Bookings without application IDs go directly to the Cancellation page.
We believe bookings without applicationIds should be very rare but users still need to be able
to cancel such bookings.
We can't go through the 'Withdrawables' flow as we can't call the the withdrawables endpoint with
an application ID as it doesn't exist. Instead we navigate directly to the 'new' cancellation
page as we did before https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/1443