Always show the Withdrawal link on the Application dashboard
Previously we only showed the 'Withdraw' link for certain application statuses.
We now want to show the link always and handover the responsibility of whether something
is withdrawable to the API (via the applications/${applicationId}/withdrawables) endpoint.
I've also renamed the 'createWithdrawElement' to 'withdrawCell' to be consistent with
other tables we generate.