Migrate from taskByType to tasks endpoint
The tasks endpoint allows for filtering by type so we can use it for this controller too.
The taskByType endpoint previously returned the tasks filtered by the user's ID so we need
to pass the new endpoint the users ID to preserve the behaviour.
There is a seperate stub 'stubTasks' which uses the same endpoint but I have opted to not remove it
in this PR as it is in the match.cy tests which are currently not in use.