Remove applications from selectable withdrawables
Since applications were added to the withdrawable.type enum they can be passed into the
withdrawableRadioOptions function.
It doesn't make sense to show them when selecting a withdrawable to withdraw as I did in [#1477](https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/1477)
so I reverted it in [#1487](https://github.com/ministryofjustice/hmpps-approved-premises-ui/pull/1487)
but this meant that the new withdrawals journey errored.
Now, we filter out the application from the list of possible withdrawables which should prevent
the error and prevent applications showing up as a selectable withdrawable.