Only show the selected withdrawable type
Following #1491 #1487 this commit ensures that the 'Select your placement'/'Select your placement request' screens only show the withdrawables of the type the user has shown on the preceding screen.
Previously we would show all the possible withdrawables. Now if a use selects 'placement' on the 'What do you want to withdraw?' screen they will only see placements on the following screen.
This slipped through the net as we were only testing that the withdrawables we wanted to see were in the view but now we also test that they withdrawables that shouldn't be visible aren't there also.
As part of this work I have updated some of the intergration test descriptions and removed
the unused 'selectedWithdrawableType' variable from the withdrawables controller's res.render
call.