Sort withdrawables by date
It makes sense that users see the withdrawables in date order.
This is complicated by the fact a withdrawable could have multiple start and end dates.
Lets use the first date, if there is one. There should always be one but just in case there
isn't we will throw an error. The only type of withdrawable that won't have a start date is an application.
There should only ever be one application so if we have reached this screen with an application then an
error will be thrown.