Change heading order
The accessibility tests were failing once the test in the following commit was added
(asserting that the 'Create placement request' button does not show for withdrawn applications).
In order to fix this I have added a H2 so there isn't a jump from h1 to h3. I have made H1
the main heading of the page and the H2 more specific and supplementary information to the H1.