Run e2es in parallel
This follows the instructions in:

https://playwright.dev/docs/ci#circleci

To avoid stamping on user’s permissions, we also have a user for each
shard, with the login details stored as with a prefix
(so `HMPPS_AUTH_USERNAME_1, HMPPS_AUTH_PASSWORD_1 etc etc), we then get
the auth details for each shard in the setup process with a bit of
Bash jiggery pokery and then run a selection of tests on each shard.
This reduces the total runtime to around 3 minutes!