Add details of placement request that has been removed to flash
We want the flash to include details of the Placement Request in order for the user to know
with confidence that they have removed the correct Placement Request.
We have made the confirmation message more generic here so the date isn't taken into account, this
content is tested in the controller instead.