Add detail to placement application withdrawal message
Now that the API returns placement dates we can use them to be more specific about what has been
withdrawn.
We fall back to a less specific message in case of error.
It makes sense to use the same message for both placement requests and requests for placement so I have extracted
it to a util function.
Having the same message show for PAs and PRs allows us to simplify the integration tests a little.
However we can no longer assert on the full withdrawal message so we just check that a message
shows that contains 'Request for placement'