Use placementDates from placementApplication object if it exists
If a placement application contains multiple dates on submission a placement application review is created
for each of these dates. Therefore the source of truth post-submission is no longer in the 'data blob'
but the placementDates key on the placementApplication object. Whilst we are moving to a model where there
is only one date per placement application review we still need to be able to handle multiple dates for
backward compatibility.
The test asserts that wherever the data comes from it looks the same to the user