Only show the withdraw button to users who made the application
Previously we gave the option to all users viewing an application to withdraw it.
However the API only allows the user who made the application to withdraw it so when the user
reached the 'what would you like to withdraw?' screen they weren't seeing any options to choose.
This change hides the button so they can't reach that screen. The clientside JS will break if
the button isn't in the DOM however, so we check the button exists before initialising it.