Use the dates from placementApplication for if they exist
When showing the mini-assessment we also want to make sure we display the placementDates
from the placement application if they exist, otherwise we fall back to the dates from the
document blob which we used previously.
There's a lot of DRYing up to be done in this file but time constraints make that difficult
currently.