Use duration and expectedArrivalDate returned from the API
Previously the API didn't return the PlacementRequestDetail when the /placement-requests/:id/withdrawal
endpoint was called.
This meant we had to do a GET in the 'new' method of the withdrawalsController and pass
the duration and expectedArrivalDate through the view as a hidden input and back into the create method.
Sometimes however a use could reach the 'new' method without having permission to call the
GET endpoint which would cause an error.
Now we remove the GET and use the PlacementRequestDetail that the API now returns to render the
details in the flash message