Add input for users to describe the 'other' reason
We previously had this input so are just adding it back it.
The conditional value makes it difficult to add using the 'convertKeyValuePairsToRadioItems' functions
so we have to add it manually.
For most of our forms we perform validation in the API but to introduce this here we would
have to ensure the UI and API were deployed to different environments at the same time. Under
current deployment conditions this is difficult so we're performing the validation in the UI
instead.
This can be moved back to the API when we are deploying more freely again.