Remove unset dates when updating the body
We’ve had issues where a user inputs a date and then unchecks a
checkbox. If the date is invalid, we don’t validate when updating
the data, so when we return the response, we get an error. We need
to ensure that we filter out any date inputs that have not been
explicitly selected.