Dry up Apply pages that use UserDetails
These pages collect the same pieces of information about different people (EG the applicant and case manager when they're different people).
In order to ensure when one page is updated the other is too they should share as much of the code
that defines these properties as possible.
This also necesitates changes to the schema.