Show case manager information in asssess
Now we have the ability to retrieve the case manager information directly from the
application object we should do that.
But we need to be able to support legacy applications which many not have this information.
This complicates things so I have extracted the functionality out in to two helper functions to improve readability.