Filter withdrawal reasons by role
Only CRU users should see capacity-related reasons when withdrawing a
placement. Non-CRU users should see all other reasons

This also removes `selectedReason` as a parameter from
`placementApplicationWithdrawalReasons`. The only places this method was
being used previously were two view templates, and in those
`selectedReason` was being passed to the method but not the the views
themselves, so it would have been `undefined`

Given the two form pages for which this method is used to generate
content are on `new` controller actions, I'm making an assumption here
that there won't be any existing data to prepopulate as the
`selectedReason`, so I'm removing it rather than looking for any
existing data with which to prepopulate the field