Add back button to placement request reasons page
We need to show a back button to users on this page.
The link necessitates an application ID which we need to pass through via flash (rather than
adding another request to GET the placement app). As we can't guarantee the link will always be in the flash and the back
button will break without it we only show the back button if it exists