Simplify placement criteria objects/types
This was pretty complicated, and we were doing things like getting an
array of the keys of an object whose keys are made up of an array of
strings, rather than just using said array of strings. This simplifies
objects and types and attempts to improve naming clarity

I've also removed `placementCriteriaUtils.test.ts`, since
`placementCriteriaUtils.ts` contains no functions: it's just a bunch of
objects and types. Utils might be slightly misleading in the name

In `matchUtils.test.ts`, because we're mocking `./utils.ts` and
`placementCriteriaUtils.ts` now uses the `filterByType` function from
there, the labels used in `groupedCriteria` were coming back as
`undefined`. I tried using `when`, but the labels still needed importing
and were undefined on import. Instead I've just mocked the relevant
objects from `placementCriteriaUtils.ts`