Add conditionally shown 'other' notes input
When the user selects 'other' we want them to have to describe their reasoning for cancelling/withdrawing the placement.
I have added validation so they must supply a description if they select 'other'.
Also as the 'Other' option is added in the UI and not supplied by the API the API may not recognise it
as a valid reason. Therefore if 'other' is selected we remove the reason when sending the request to the API
and only send the date and the notes.