Add util for Cancellation reasons
We need to add a text box for users to describe why they are picking the 'other' reason.
This is made more difficult by the reasons being supplied by the API.
Firstly we loop through the reasons to get them in the right shape for the inputs in the view.
Then we add the conditional.