WIP Use AP type enum/union in Apply
TODO: once API is updated, use new OpenAPI types and remove temporary
types. Might also need to update the use of the AP type enum in other
places where 'normal' is expected instead of 'standard' if the current
`ApType` is what gets updated. If lucky, might be able to simplify the
number of AP type types there are

Some of the PIPE-specific testing that wasn't done for ESAP is removed
here, since PIPE, ESAP, and standard applications effectively become
equal, and are just indicated via a single enum string field