Add util for Cancellation reasons
We need to add a text box for users to describe why they are picking
the 'other' reason. We do this by converting the `cancellationReasons`
to an array of radio items, then adding a `conditional` based on if the
text of that item is `Other`. It’s a bit of a hack, but unavoidable
because of the way we handle reference data.