Add conditionally shown 'other' notes input
When the user selects 'other' we want them to have to describe their
reasoning for cancelling/withdrawing the placement.

The validation for this is handled by the API, so we just need to add
some integration tests here. I’ve added a trait to the newCancellation
factory, so we have one source of truth for what the `Other` option is.