Add Placement Applications to the assessment dashboard
While this may make sense to the user it's confusing two different entities in the code.
So if the user clicks the 'Requests for placement' tab we need to call a different endpoint.
I think it's clearest to have to seperate calls to res.render rather than have lots of booleans
flying about.
The new Placement Applications dashboard page doesn't use sorting currently so I have made the
sorting fields nullable when mocking, otherwise we wouldn't be able to use the stubGetAllTasks
function.