Add a simple client/service to get a list of users
To filter by allocated user, we need a list of all users with specific
roles. By default, the `getUsers` Client/Service adds a pageNumber,
which enables pagination on the user side. To get around this and avoid
overcomplicating the existing service and client methods, we create
a simple `getUserList` service and client method to just return all
users with the roles we need.