Add startDateOutsideOfNationalStandardsTimescales
This is a helper that returns if an application was submitted with less
than six months notice. date-fns doesn’t allow us to get the difference
in months in a predictable way that takes into account fractional months
so subtract six months from the arrivalDate and then check if it is
before today. If it is, then we know it’s outside the timescales.