Verify feedback form works
It would be good to know next time the feedback form is changed.
I think it makes sense to live in the integration test code even though it is
more of an 'e2e' test so we are alerted quicker if it breaks.
We need to navigate to the page and test it loads rather than just verify the link is what we'd expect as
the link may stop working without us knowing.