Show both the applicant and the case manager details on the assess screen
We need to the user the details for both the case manager and the applicant.
Often this is the same person so we only show one set of details and indicate in the heading that they're
the same person. Otherwise we show both sets of details.
As the backfill has now been completed in the API to fill the 'caseManagerIsNotApplicant', 'applicantUserDetails'
and 'caseManagerUserDetails' we can remove the code responsible for fetching this information from the application
itself and just use the first class properties instead.