Add `BackwardsCompatibleApplyApType` type
Currently the AP type question in Apply can have a value of 'standard'.
An incoming change will shift this question to use the `ApType` type,
which instead uses 'normal'. This adds an AP type type that includes
'standard' so that we can write code that is backwards compatible with
application data blobs that predate this change