Correct ordering of AP types
The ordering of the AP types is significant to users.
It's important that it is the same on the ApType page as it is on the MatchingInformation page.
Any change to one should be reflected in the other.
In order to achieve this I have had to switch to an array for the AP types as objects can't guarantee that
their ordering will be preserved. We then use the same array in both places but we had to map it to our needs
in each usage. This is awkward but preferable to them being too loosely coupled.