Add a class to render a summary from document
When viewing applications in the read only version, or when assessings,
we still build up the response using the `data` response and the form
classes. This causes issues if the schema changes, as we will get
errors. This was one of the reasons why we have a persisted `document`
field in an application, but this wasn’t getting used for some reason

This adds a new `SumbmittedApplicationSummaryCards` class to get
all the sections for a form (unfortuately this is not yet persisted)
and walk through each task and page and return each question in
each page as a summary list. This is done in other places for
“Check your answers” style steps, but uses the `data` field, instead of
the document.

If an `assesmentId` is given, this will be for the assessment review
page, so we’ll want to call the `getActionsForTaskId` helper to
add some extra actions to view prison notes and Oasys responses.

There’s also a slight hack for documents, where unfortunately we have
to reach into the `data` field, but hopefully this should be fairly
static.

I added another `sections` argument to the contructor to make testing
easier. It defaults to `Apply.sections.slice(0,1)` which is all sections
apart from Check Your Answers