Handle offline applications in summary cards
The recently added `SubmittedApplicationSummaryCards` class is throwing
errors for offline applications. Application documents are a JSON blob
representing the responses to the online form at submission time.
Offline applications by definition don't have these. The summary cards
class was attempting to read from the document without checking it
exists (strict mode might have caught this)

This fixes that issue, adds a related unit test, and makes the related
integration test setup more realistic so that it can catch this error if
it reappears