Prevent infinite loop in sufficient-information/sufficient-information-confirm
An infinite loop can occur between the Sufficient Information &
Sufficient Information Confirm pages if the user answers 'no' to the
Sufficient Information question (stating there is not sufficient
information to make a decision) and 'no' to the confirmation question
(stating the user is not sure they want to request more information).
To prevent this if the user answers 'no' there is not sufficient
information to make a decision AND 'no' they are not sure they want to
request further information then we delete their response to the
confirmation question so that when they reach the confirmation page the
application is not in an infinite loop.