Add feature flag to toggle whether users see confirmation page
We want to see if how the user journey plays out without the
confirmation page.
The easiest way to do this is with a feature flag toggled on in dev and
off in other environments.
We need to add the private method previouslyRequested to follow the
logic present in alreadyConfirmed method on the
SufficientInformationConfirm page. This prevents the
clarificationNote/query being submitted multiple times.