Pass feature flags to Tasklist Page constructor
When we create a new Page class it needs to be able to access the
feature flags which neccessitates an async API call.
Normally async work for the Page classes is done in the class'
'initialize' method but this method isn't called in the Tasklist
Service.
So we can either:
- call the 'initialize' method on the Page classes in the Tasklist
Service. This seems like overkill to just fetch the feature flags as it
could result in many, many API calls
- fetch the feature flags and pass them into the Page class'
constructor. This involves modifying the code in more places but less
API calls and allows us to set the flags value in the integration tests
more easily.
Another option that I discounted was using middleware to fetch the flags
but this required as many code changes with the performance impact of
fetching the flags on every request OR implementing some caching
mechanism that introduces extra complexity.