Remove `getActingUser` HMPPS Auth call
This has been retired in favour of the manage users API. However,
all of the profile information we need is fetched via the Approved
Premises API, so this is a call that we no longer need to make.

I had to do a bit of jiggery pokery with the task integration
tests, because they assumed a user wouldn’t have an AP Area
by default, which is no longer the case anyway.