Add getAll method to feature flag service
Even though we only have one flag at the moment we will, in future
commits, want to retrieve all the flags.
This method does that. There will be a performance hit if there are a
large amount of flags which is one more reason that we need to be
concious to not add any more flags than neccessary and remove flags as
soon as possible