Add retrieveFeatureFlag util
Env vars can only be saved as strings which means boolean values will
come back as 'true' or 'false' which are both truthy values. In order to
avoid repeating logic throughout the application we can use this util.
It will also come in handy when we want to mock feature flags