Add timeline pages
We add a couple of pages, the first one is to enter the CRN the user
wishes to view the timeline for and the second is to view the timeline.
There are no mocks or copy for these pages yet so I've had to improvise.
As this is a new section of the service (people) there is more
boilerplate than usual.