Move people controller
The timeline stands seperate from the other sections of the service
(Apply, Assess, Manage etc.) as it needs to be accessible by all users.
We will create a new section for it called 'People'. It makes sense that
the people controller sits alongside it, so lets move it there to begin
with.