Extract timeline macro
We will need this timeline in multiple places so lets extract it into a
macro

As we do this lets move the method to verify the timeline is shown
correctly up to the Page class so we can use it in multiple places too

Now we have both a Personal timeline and an Application timeline we need
to clarify which this function deals with.
I also delete it as an explicitly defined nunjucks function as it is
already available in the ApplyUtils object
I have also added some code to handle when there isn't an 'occurredAt'
property on a timeline event. It doesn't make sense to have a timeline
event without an occurredAt propety but the spec says it's possible and
it's relatively easy to handle.